build: Make json, msgpack and avro optional dependencies
These are only needed if the user is using this strategy and a particular
encoding format, and are likely not needed by the majority of users.
