+++ arroyo/processing/strategies/decoder/base.py
@@ -1,15 +1,8 @@
 from __future__ import annotations
 
-from avro.io import BinaryDecoder, DatumReader
-from avro.errors import SchemaResolutionException
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import Generic, Optional, Sequence, Tuple, TypeVar
-from io import BytesIO
-
-import fastjsonschema
-import msgpack
-import rapidjson
 
 from arroyo.backends.kafka import KafkaPayload
 from arroyo.processing.strategies.abstract import ProcessingStrategy
@@ -82,61 +75,3 @@ class Codec(ABC, Generic[T]):
         If validate is true, validation is performed.
         """
         raise NotImplementedError
-
-
-class JsonCodec(Codec[object]):
-    def __init__(self, json_schema: Optional[object] = None) -> None:
-        if json_schema is not None:
-            self.__validate = fastjsonschema.compile(json_schema)
-        else:
-            self.__validate = lambda _: _
-
-    def decode(self, raw_data: bytes, validate: bool) -> object:
-        decoded = rapidjson.loads(raw_data)
-        if validate:
-            self.validate(decoded)
-        return decoded
-
-    def validate(self, data: object) -> None:
-        try:
-            self.__validate(data)
-        except Exception as exc:
-            raise ValidationError from exc
-
-
-class MsgpackCodec(Codec[object]):
-    """
-    This codec assumes the schema is json.
-    """
-
-    def __init__(self, json_schema: Optional[object] = None) -> None:
-        if json_schema is not None:
-            self.__validate = fastjsonschema.compile(json_schema)
-        else:
-            self.__validate = lambda _: _
-
-    def decode(self, raw_data: bytes, validate: bool) -> object:
-        decoded = msgpack.unpackb(raw_data)
-        if validate:
-            self.validate(decoded)
-        return decoded
-
-    def validate(self, data: object) -> None:
-        try:
-            self.__validate(data)
-        except Exception as exc:
-            raise ValidationError from exc
-
-
-class AvroCodec(Codec[object]):
-    def __init__(self, avro_schema: object) -> None:
-        self.__schema = avro_schema
-        self.__reader = DatumReader(self.__schema)
-
-    def decode(self, raw_data: bytes, validate: bool) -> object:
-        assert validate is True, "Validation cannot be skipped for AvroCodec"
-        decoder = BinaryDecoder(BytesIO(raw_data))
-        try:
-            return self.__reader.read(decoder)
-        except SchemaResolutionException as exc:
-            raise ValidationError from exc
